requirements.txt
openai==0.27.2
python-chess==1.999
python-dotenv==1.0.0